+++ library/xgpl_src/IsoAgLib/comm/Part6_VirtualTerminal_Client/impl/vtobjectstringvariable_c.cpp
@@ -151,7 +151,7 @@ vtObjectStringVariable_c::setValueRef(const char* newValue, bool b_updateObject,
 
   setStringToStream( newValue ); // use MultiSendStreamer with mpc_stringToStream set!
   uint16_t ui16_tempLen = 0;
-  if (newValue != NULL ) ui16_tempLen = (CNAMESPACE::strlen (newValue) <= get_vtObjectStringVariable_a()->length) ? CNAMESPACE::strlen (newValue) : get_vtObjectStringVariable_a()->length;
+  if (newValue != NULL ) ui16_tempLen = uint16_t( (CNAMESPACE::strlen (newValue) <= get_vtObjectStringVariable_a()->length) ? CNAMESPACE::strlen (newValue) : get_vtObjectStringVariable_a()->length );
   setStrLenToSend( ui16_tempLen );
   __IsoAgLib::getVtClientInstance4Comm().getClientByID (s_properties.clientId).commandHandler().sendCommandChangeStringValue (this, b_enableReplaceOfCmd);
 }
